Fix exception compatibility problems of ByteBuffer, CharBuffer and DoubleBuffer, also add some testcases to enforce it.
git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@423386 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/nio/src/main/java/java/nio/ReadOnlyCharArrayBuffer.java
@@ -1,4 +1,4 @@
-/* Copyright 2004 The Apache Software Foundation or its licensors, as applicable
+/* Copyright 2004, 2006 The Apache Software Foundation or its licensors, as applicable
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -89,7 +89,7 @@ final class ReadOnlyCharArrayBuffer extends CharArrayBuffer {
     }
 
     public CharBuffer put(String src, int start, int end) {
-        if ((start < 0 ) || (end < 0) || start + end > src.length()) {
+        if ((start < 0 ) || (end < 0) || (long)start + (long)end > src.length()) {
             throw new IndexOutOfBoundsException();
         }
         if(src == null){
